Move DefaultNullQgsDateEdit function to utils and simplify form handling
