+++ ee_plugin/ui/forms.py
@@ -1,33 +1,17 @@
 from typing import Optional, Callable
 from qgis import gui
-from qgis.PyQt import QtWidgets, QtCore
+from qgis.PyQt import QtWidgets
 import ee
 
 from .utils import (
     build_form_group_box,
     build_vbox_dialog,
     call_func_with_values,
+    DefaultNullQgsDateEdit,
 )
 
 from .. import Map, utils
 from ..utils import translate as _
-
-
-def DefaultNullQgsDateEdit(
-    *, date: Optional[QtCore.QDate] = None, displayFormat="yyyy-MM-dd", **kwargs
-) -> gui.QgsDateEdit:
-    """Build a QgsDateEdit widget, with null default capability."""
-    # NOTE: Specifying a displayFormat guarantees that the date will be formatted as
-    # expected across different runtime environments.
-    d = gui.QgsDateEdit(**kwargs, displayFormat=displayFormat)
-    # NOTE: It would be great to remove this helper and just use the built-in QgsDateEdit
-    # class but at this time it's not clear how to make a DateEdit widget that initializes
-    # with a null value. This is a workaround.
-    if date is None:
-        d.clear()
-    else:
-        d.setDate(date)
-    return d
 
 
 def add_feature_collection_form(
